Ejercicio_Practico_01/requirements.txt
Flask==3.0.2
flask-jwt-extended==4.6.0
python-dotenv==1.0.1
requests==2.32.3
num2words==0.5.14
flasgger==0.9.7.1
google-cloud-translate==3.24.0